làm lại

+++ src/main/java/poly/cinema/entity/SuatChieu.java
@@ -27,7 +27,7 @@ public class SuatChieu {
     private String maPhong;
     private LocalDate ngayChieu;
     private LocalTime gioChieu;
-    private int giaVe;
+    private double giaVe;
 
     private PhongChieu phong; 
 }
